+++ content/changelog/2023-09-08-console.md
@@ -9,7 +9,7 @@ In addition, the **Compute** widget on the Neon **Dashboard** now shows the **De
 
 ### Fixes & improvements
 
-- API: Neon API v1 deprecation was [announced in December, 2022](/docs/release-notes/2022-12-28-console). We have removed code for this early version of the Neon API. If you have not yet migrated your applications to the current version of the [Neon API](https://api-docs.neon.tech/reference/getting-started-with-neon-api), please do so now.
+- API: Neon API v1 deprecation was [announced in December, 2022](/docs/changelog/2022-12-28-console). We have removed code for this early version of the Neon API. If you have not yet migrated your applications to the current version of the [Neon API](https://api-docs.neon.tech/reference/getting-started-with-neon-api), please do so now.
 - UI: Added a **Billing** page to the Neon Console for Neon Enterprise and Platform Partnership custom plan users. The **Billing** page is accessible from the sidebar in the [Neon Console](https://console.neon.tech/app/projects).
 - UI: Addressed usability issues on the **Custom** tab of the **Pro Plan Cost Estimator**, which is accessible from the Free Tier **Billing** page in the Neon Console. Percentage values were not restored when switching from **Enter your own values** and **Use a percentage**. Additionally, minimum, maximum, and increment values were adjusted for **Compute** and **Project storage** metrics.
 - UI: The billing period drop-down menu at the top right corner of the **Billing** page now uses the current billing period as the default selection instead of the oldest billing period. This change applies to Neon Pro plan accounts.
