+++ src/components/pages/changelog/changelog-filter/index.js
@@ -1,0 +1,1 @@
+export { default, CHANGELOG_CATEGORIES } from './changelog-filter';
