+++ src/components/pages/changelog/changelog-list/index.js
@@ -1,0 +1,1 @@
+export { default } from './changelog-list';
