--- src/components/pages/release-notes/release-note-list/index.js
@@ -1,1 +1,0 @@
-export { default } from './release-note-list';
