--- src/components/pages/release-notes/release-notes-filter/index.js
@@ -1,1 +1,0 @@
-export { default, RELEASE_NOTES_CATEGORIES } from './release-notes-filter';
